feat(driver): 实现 ED047TC1 电子墨水屏驱动及波形优化
- 新增 ED047TC1 驱动模块，支持 4.7 寸 16 级灰度电子墨水屏
- 实现帧缓冲区管理，支持像素读写和矩形区域绘制
- 提供多种绘图方法，包括画线、矩形及像素清屏
- 集成波形温度自适应，支持不同温度对应的刷新波形
- 定义驱动所需引脚配置及默认引脚映射
- 完善错误处理，支持多种驱动异常状态
- 提供显示状态管理及绘图模式切换接口
- 优化波形模块温度变量类型并提升接口一致性
- 添加丰富的使用示例文档和注释说明
- 调整模块导出接口，便于外部调用和扩展

+++ src/driver/ed047tc1/waveform.rs
@@ -155,9 +155,9 @@ pub struct WaveformMode {
 #[derive(Debug, Clone, Copy)]
 pub struct WaveformTemperatureInterval {
     /// 最低温度（摄氏度）
-    pub min: i32,
+    pub min: i8,
     /// 最高温度（摄氏度）
-    pub max: i32,
+    pub max: i8,
 }
 
 // ============================================================================
@@ -179,13 +179,10 @@ pub struct WaveformTemperatureInterval {
 pub struct Waveform {
     /// 可用模式数量
     pub number_mode: u8,
-
     /// 温度范围数量
     pub number_temperature_range: u8,
-
     /// 所有可用的波形模式
     pub mode_data: &'static [&'static WaveformMode],
-
     /// 温度区间配置
     pub temperature_interval: &'static [WaveformTemperatureInterval],
 }
@@ -230,7 +227,7 @@ impl Waveform {
     ///
     /// # 说明
     /// 如果没有匹配的温度范围，会回退到第一个范围
-    pub fn get_phases_for_temperature(&self, mode: &WaveformMode, temperature: i32) -> Option<&'static WaveformPhase> {
+    pub fn get_phases_for_temperature(&self, mode: &WaveformMode, temperature: i8) -> Option<&'static WaveformPhase> {
         // 遍历温度区间，找到匹配的范围
         for (i, interval) in self.temperature_interval.iter().enumerate() {
             if temperature >= interval.min && temperature <= interval.max {
@@ -537,7 +534,7 @@ static ED047TC1_MODE17: WaveformMode = WaveformMode {
 // ============================================================================
 
 /// ED047TC1 的温度区间配置
-/// 当前只配置了一个温度范围：20-30°C（室温）
+/// 当前只配置了一个温度范围：20 - 30°C（室温）
 static ED047TC1_TEMPERATURE_INTERVAL: [WaveformTemperatureInterval; 1] = [WaveformTemperatureInterval { min: 20, max: 30 }];
 
 /// ED047TC1 所有可用的波形模式
